// src/app/models/vehicle.model.ts
export class vehicle {
    id?: string;
    userId?: string;
    userDescription?: string;
    vehicleCategoryId?: string;
    vehicleCategoryDescription?: string;
    vehicleTypeId?: string;
    vehicleTypeDescription?: string;
    description?: string;
    registrationNumber?: string;
    maxLoadWeight?: number;
    maxLoadHeight?: number;
    maxLoadWidth?: number;
    maxLoadLength?: number;
    maxLoadVolume?: number;
    availableCapacity?: number;
    availableFrom?: Date;
    availableTo?: Date;
    originatingAddressLabel?: string;
    originatingAddressLat?: number;
    originatingAddressLon?: number;
    destinationAddressLabel?: string;
    destinationAddressLat?: number;
    destinationAddressLon?: number;
    avatar?: string;
    statusId?: string;
    statusDescription?: string;
}